# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server Petstore server.",
        "title": "Swagger Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost",
    "basePath": "/",
    "paths": {
        "/api/postInst": {
            "post": {
                "description": "get struct array by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "operationId": "get-struct-array2-by-string",
                "parameters": [
                    {
                        "description": "request job",
                        "name": "loginUserRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/router.InstructionD"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/view/ApplicationDB": {
            "get": {
                "description": "Retrieves application database records by UUID, or all records if no UUID is provided",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ApplicationDB"
                ],
                "summary": "Get application database records",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Application UUID",
                        "name": "uuid",
                        "in": "query"
                    }
                ],
                "responses": {}
            }
        },
        "/view/JobDataDB": {
            "get": {
                "description": "Retrieves all job data from the database",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "JobDataDB"
                ],
                "summary": "Get all job data",
                "responses": {}
            }
        },
        "/view/OperationLogDB": {
            "get": {
                "description": "Retrieves all operation logs from the database",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OperationLogDB"
                ],
                "summary": "Get all operation logs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Agent UUID",
                        "name": "uuid",
                        "in": "query"
                    }
                ],
                "responses": {}
            }
        },
        "/view/SystemInfoDB": {
            "get": {
                "description": "Retrieves system information records by UUID, or all records if no UUID is provided",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SystemInfoDB"
                ],
                "summary": "Get system information records",
                "parameters": [
                    {
                        "type": "string",
                        "description": "System UUID",
                        "name": "uuid",
                        "in": "query"
                    }
                ],
                "responses": {}
            }
        },
        "/view/agentStatus": {
            "get": {
                "description": "Retrieves agent status by UUID, or all agents if no UUID is provided",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AgentStatus"
                ],
                "summary": "Get agent status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Agent UUID",
                        "name": "uuid",
                        "in": "query"
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
        "router.InstructionD": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string",
                    "default": "ExecutePayLoad"
                },
                "agentUUID": {
                    "type": "string",
                    "default": "09a4e53c7a1c4b4e9a519f36df29d8a2"
                },
                "procedureID": {
                    "type": "string",
                    "default": "P_DefenseEvasion_Kimsuky_001"
                }
            }
        }
    }
}